Add arity protocols
